Move `API\Addresses\*` to `API\Model\Wallet`
