+++ includes/action-scheduler/interface-background-jobs-scheduler-interface.php
@@ -13,7 +13,7 @@
 
 namespace BrianHenryIE\WP_Bitcoin_Gateway\Action_Scheduler;
 
-use BrianHenryIE\WP_Bitcoin_Gateway\API\Addresses\Bitcoin_Wallet;
+use BrianHenryIE\WP_Bitcoin_Gateway\API\Model\Wallet\Bitcoin_Wallet;
 use DateTimeInterface;
 
 interface Background_Jobs_Scheduler_Interface {
