+++ includes/api/model/wallet/enum-bitcoin-address-status.php
@@ -3,7 +3,7 @@
  * @package    brianhenryie/bh-wp-bitcoin-gateway
  */
 
-namespace BrianHenryIE\WP_Bitcoin_Gateway\API\Addresses;
+namespace BrianHenryIE\WP_Bitcoin_Gateway\API\Model\Wallet;
 
 enum Bitcoin_Address_Status: string {
