+++ includes/api/model/wallet/interface-bitcoin-wallet-interface.php
@@ -5,7 +5,7 @@
  * @package    brianhenryie/bh-wp-bitcoin-gateway
  */
 
-namespace BrianHenryIE\WP_Bitcoin_Gateway\API\Addresses;
+namespace BrianHenryIE\WP_Bitcoin_Gateway\API\Model\Wallet;
 
 interface Bitcoin_Wallet_Interface {
