+++ includes/api/services/results/class-get-wallet-for-xpub-service-result.php
@@ -8,7 +8,7 @@
 
 namespace BrianHenryIE\WP_Bitcoin_Gateway\API\Services\Results;
 
-use BrianHenryIE\WP_Bitcoin_Gateway\API\Addresses\Bitcoin_Wallet;
+use BrianHenryIE\WP_Bitcoin_Gateway\API\Model\Wallet\Bitcoin_Wallet;
 use BrianHenryIE\WP_Bitcoin_Gateway\API\Services\Bitcoin_Wallet_Service;
 
 /**
